+++ src/graphsdl.h
@@ -62,7 +62,7 @@ extern int32 os_readpalette(int32 colour, int32 mode);
 extern void sdl_screensave(char *fname);
 extern void sdl_screenload(char *fname);
 extern void swi_swap16palette(void);
-extern int32 readmodevariable(int32 scrmode, int32 var);
+extern size_t readmodevariable(int32 scrmode, int32 var);
 extern void screencopy(int32 src, int32 dst);
 extern int32 get_maxbanks(void);
 extern void refresh_location(uint32 offset);
